+++ app/domain/booking_states/awaiting_tenant.rb
@@ -34,10 +34,9 @@ module BookingStates
     end
 
     after_transition do |booking|
-      booking.deadline&.clear
-      booking.deadlines.create(length: booking.organisation.settings.awaiting_tenant_deadline,
-                               postponable_for: booking.organisation.settings.deadline_postponable_for,
-                               remarks: booking.booking_state.t(:label))
+      booking.set_deadline(length: booking.organisation.settings.awaiting_tenant_deadline,
+                           postponable_for: booking.organisation.settings.deadline_postponable_for,
+                           remarks: booking.booking_state.t(:label))
       MailTemplate.use!(:awaiting_tenant_notification, booking, to: :tenant, &:autodeliver!)
       next if booking.agent_booking.blank?