+++ app/domain/booking_states/provisional_request.rb
@@ -27,10 +27,9 @@ module BookingStates
     end
 
     after_transition do |booking|
-      booking.deadline&.clear
-      booking.deadlines.create(length: booking.organisation.settings.provisional_request_deadline,
-                               postponable_for: booking.organisation.settings.deadline_postponable_for,
-                               remarks: booking.booking_state.t(:label))
+      booking.set_deadline(length: booking.organisation.settings.provisional_request_deadline,
+                           postponable_for: booking.organisation.settings.deadline_postponable_for,
+                           remarks: booking.booking_state.t(:label))
       booking.tentative!
       next if booking.committed_request