+++ app/models/booking_conditions/tenant_attribute.rb
@@ -1,0 +1,50 @@
+# frozen_string_literal: true
+
+# == Schema Information
+#
+# Table name: booking_conditions
+#
+#  id                :bigint           not null, primary key
+#  compare_attribute :string
+#  compare_operator  :string
+#  compare_value     :string
+#  group             :string
+#  must_condition    :boolean          default(TRUE)
+#  qualifiable_type  :string
+#  type              :string
+#  created_at        :datetime         not null
+#  updated_at        :datetime         not null
+#  organisation_id   :bigint
+#  qualifiable_id    :bigint
+#
+# Indexes
+#
+#  index_booking_conditions_on_organisation_id                      (organisation_id)
+#  index_booking_conditions_on_qualifiable                          (qualifiable_id,qualifiable_type,group)
+#  index_booking_conditions_on_qualifiable_id_and_qualifiable_type  (qualifiable_id,qualifiable_type)
+#
+# Foreign Keys
+#
+#  fk_rails_...  (organisation_id => organisations.id)
+#
+
+module BookingConditions
+  class TenantAttribute < BookingCondition
+    BookingCondition.register_subtype self
+
+    attribute :compare_operator, default: -> { :'=' }
+
+    compare_attribute country_code: ->(tenant:) { tenant.country_code&.upcase }
+
+    compare_operator(**NUMERIC_OPERATORS)
+
+    validates :compare_attribute, :compare_operator, presence: true
+
+    def evaluate!(booking)
+      actual_value = evaluate_attribute(compare_attribute, with: { tenant: booking.tenant })
+      return if actual_value.blank? || compare_value.blank?
+
+      evaluate_operator(compare_operator, with: { actual_value:, compare_value: })
+    end
+  end
+end