+++ db/migrate/20241125082419_remove_deadline_id_for_bookings.rb
@@ -1,0 +1,11 @@
+class RemoveDeadlineIdForBookings < ActiveRecord::Migration[7.2]
+  def change
+    reversible do |direction|
+      direction.up do
+        Deadline.where.not(id: Booking.pluck(:deadline_id)).delete_all
+      end
+    end
+
+    remove_column :bookings, :deadline_id, :bigint, index: true
+  end
+end