+++ spec/models/deadline_spec.rb
@@ -42,7 +42,7 @@ RSpec.describe Deadline, type: :model do
   end
 
   describe '#create' do
-    subject(:deadline) { booking.deadlines.create(length:) }
+    subject(:deadline) { booking.set_deadline(length:) }
 
     context 'with length set' do
       let(:length) { 1.week }