+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 24.11.1
+
+Released 25.11.2024
+
+- Fix: Refactor deadlines to fix out-of-order issues
+- Fix: Order Booking Question Responses correctly
+- Feature: Add Tenant Attributes as BookingCondition
+
 ## 24.10.2
 
 Released 23.10.2024